% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDispVec.R
\name{myDispVec}
\alias{myDispVec}
\title{myDispVec}
\usage{
myDispVec(x, ...)
}
\arguments{
\item{x}{a list inherited from `sampleG()`}

\item{...}{passes additional arguments throughout the function}
}
\value{
a 3D plot of the displacement (residual) vectors
}
\description{
Plots the displacement (residual) vectors
}
\examples{
\dontrun{myDispVec(x = first)}
}
