% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.samg.R
\name{print.samg}
\alias{print.samg}
\title{print.samg}
\usage{
\method{print}{samg}(x = list(), ...)
}
\arguments{
\item{x}{a list inherited from `sampleG()`}

\item{...}{passes additional arguments throughout the function}
}
\value{
a plot of the `p` vectors
}
\description{
An S3 Method for `sampleG()` which plots the first three components of the displacement vectors and returns a named list
}
\examples{
\dontrun{print(x = l)}
}
