% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleG.R
\name{sampleG}
\alias{sampleG}
\title{sampleG}
\usage{
sampleG(x = data.frame())
}
\arguments{
\item{x}{a data frame}
}
\value{
a named list
}
\description{
A function which uses the quantitative data frame to produce xbar the sample mean vector and the centered matrix
}
\examples{
\dontrun{sampleG(x = ddt[1:10, c("LENGTH", "WEIGHT", "DDT")])}
}
