% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.samg.R
\name{summary.samg}
\alias{summary.samg}
\title{summary.samg}
\usage{
\method{summary}{samg}(object = list(), ...)
}
\arguments{
\item{object}{a list from `sampleG()`}

\item{...}{passes additional arguments throughout the function}
}
\value{
a named list
}
\description{
An S3 Method for `sampleG()` which returns a named list containing the generalized variance, total variance, sample cov matrix S, and the eigenvalues and eigenvectors of S
}
\examples{
\dontrun{summary(x=l)}
}
